Make each game object only have a single Lua userdata
Now a table exists which stores userdata for stuff like mobjs,
players, and psprs, to allow for each of these objects to be
represented in Lua by a single userdata, instead of multiple ones
pointing to the same object. This is cleaner and also means that
an _eq metamethod is no longer necessary. rawequal now works as expected.